+++ S3Sort/selsort.py
@@ -1,11 +1,3 @@
-from random import randint as rand
-
-def listGen(n):
-    randlist = []
-    for i in range(n):
-        randlist.append(rand(1,100))
-    return randlist
-
 def findSmallest(sortlist):
     smallest = sortlist[0]
     smallest_index = 0
@@ -21,8 +13,3 @@ def selectionSort(sortlist):
         smallest = findSmallest(sortlist)
         newsortlist.append(sortlist.pop(smallest))
     return newsortlist
-
-
-newlist = listGen(10)
-print newlist
-print selectionSort(newlist)
